弾発射音がコントローラー画面で流れるように、hooksに分割
