+++ client/src/pages/@hooks/useWindowSize.ts
@@ -1,0 +1,22 @@
+import { useEffect, useState } from 'react';
+
+export const useWindowSize = () => {
+  const [windowsize, setWindowsize] = useState<{ width: number; height: number }>({
+    width: window.innerWidth,
+    height: window.innerHeight,
+  });
+
+  useEffect(() => {
+    const handleResize = () => {
+      setWindowsize({
+        width: window.innerWidth,
+        height: window.innerHeight,
+      });
+    };
+    window.addEventListener('resize', handleResize);
+    return () => {
+      window.removeEventListener('resize', handleResize);
+    };
+  }, []);
+  return windowsize;
+};
